refactor: move component types/opt from types.ts to their modules
