+++ src/components/draw/opacity.ts
@@ -1,15 +1,29 @@
-import { getInternalContext, getKaboomContext } from "../../kaboom";
-import type { OpacityComp, TweenController } from "../../types";
+import { getKaboomContext } from "../../kaboom";
+import type { Comp, EaseFunc, TweenController } from "../../types";
+
+/**
+ * The {@link opacity `opacity()`} component.
+ *
+ * @group Components
+ */
+export interface OpacityComp extends Comp {
+    /** Opacity of the current object. */
+    opacity: number;
+    /** Fade in at the start. */
+    fadeIn(time?: number, easeFunc?: EaseFunc): TweenController;
+    /** Fade out at the start. */
+    fadeOut(time?: number, easeFunc?: EaseFunc): TweenController;
+}
 
 export function opacity(a: number): OpacityComp {
     const k = getKaboomContext(this);
-    const internal = getInternalContext(k);
+    const { toFixed } = k._k;
 
     return {
         id: "opacity",
         opacity: a ?? 1,
         inspect() {
-            return `${internal.toFixed(this.opacity, 1)}`;
+            return `${toFixed(this.opacity, 1)}`;
         },
         fadeIn(time = 1, easeFunc = k.easings.linear): TweenController {
             return k.tween(
