+++ src/components/draw/polygon.ts
@@ -1,10 +1,53 @@
-import { getInternalContext, getKaboomContext } from "../../kaboom";
-import { Polygon } from "../../math";
-import type { GameObj, PolygonComp, PolygonCompOpt, Vec2 } from "../../types";
+import type { Texture } from "../../gfx";
+import { getKaboomContext } from "../../kaboom";
+import { type Color, Polygon } from "../../math";
+import type { Comp, DrawPolygonOpt, GameObj, Vec2 } from "../../types";
+
+/**
+ * The {@link polygon `polygon()`} component.
+ *
+ * @since v3001.0
+ * @group Components
+ */
+export interface PolygonComp extends Comp {
+    draw: Comp["draw"];
+    /**
+     * Points in the polygon.
+     */
+    pts: Vec2[];
+    /**
+     * The radius of each corner.
+     */
+    radius?: number | number[];
+    /**
+     * The color of each vertex.
+     */
+    colors?: Color[];
+    /**
+     * The uv of each vertex.
+     *
+     * @since v3001.0
+     */
+    uv?: Vec2[];
+    /**
+     * The texture used when uv coordinates are present.
+     *
+     * @since v3001.0
+     */
+    tex?: Texture;
+    renderArea(): Polygon;
+}
+
+/**
+ * Options for the {@link polygon `polygon()`} component.
+ *
+ * @group Components
+ */
+export type PolygonCompOpt = Omit<DrawPolygonOpt, "pts">;
 
 export function polygon(pts: Vec2[], opt: PolygonCompOpt = {}): PolygonComp {
     const k = getKaboomContext(this);
-    const internal = getInternalContext(k);
+    const { getRenderProps } = k._k;
 
     if (pts.length < 3) {
         throw new Error(
@@ -19,7 +62,7 @@ export function polygon(pts: Vec2[], opt: PolygonCompOpt = {}): PolygonComp {
         tex: opt.tex,
         radius: opt.radius,
         draw(this: GameObj<PolygonComp>) {
-            k.drawPolygon(Object.assign(internal.getRenderProps(this), {
+            k.drawPolygon(Object.assign(getRenderProps(this), {
                 pts: this.pts,
                 colors: this.colors,
                 uv: this.uv,
