+++ src/components/draw/rect.ts
@@ -1,10 +1,51 @@
-import { getInternalContext, getKaboomContext } from "../../kaboom";
+import { getKaboomContext } from "../../kaboom";
 import { Rect, vec2 } from "../../math";
-import type { GameObj, RectComp, RectCompOpt } from "../../types";
+import type { Comp, GameObj } from "../../types";
+
+/**
+ * The {@link rect `rect()`} component.
+ *
+ * @group Components
+ */
+export interface RectComp extends Comp {
+    draw: Comp["draw"];
+    /**
+     * Width of rectangle.
+     */
+    width: number;
+    /**
+     * Height of rectangle.
+     */
+    height: number;
+    /**
+     * The radius of each corner.
+     */
+    radius?: number;
+    /**
+     * @since v3000.0
+     */
+    renderArea(): Rect;
+}
+
+/**
+ * Options for the {@link rect `rect()`} component.
+ *
+ * @group Components
+ */
+export interface RectCompOpt {
+    /**
+     * Radius of the rectangle corners.
+     */
+    radius?: number;
+    /**
+     * If fill the rectangle (useful if you only want to render outline with outline() component).
+     */
+    fill?: boolean;
+}
 
 export function rect(w: number, h: number, opt: RectCompOpt = {}): RectComp {
     const k = getKaboomContext(this);
-    const internal = getInternalContext(k);
+    const { getRenderProps } = k._k;
 
     return {
         id: "rect",
@@ -12,7 +53,7 @@ export function rect(w: number, h: number, opt: RectCompOpt = {}): RectComp {
         height: h,
         radius: opt.radius || 0,
         draw(this: GameObj<RectComp>) {
-            k.drawRect(Object.assign(internal.getRenderProps(this), {
+            k.drawRect(Object.assign(getRenderProps(this), {
                 width: this.width,
                 height: this.height,
                 radius: this.radius,
