+++ src/components/misc/lifespan.ts
@@ -1,11 +1,18 @@
 import easings from "../../easings";
 import { getKaboomContext } from "../../kaboom";
-import type {
-    EmptyComp,
-    GameObj,
-    LifespanCompOpt,
-    OpacityComp,
-} from "../../types";
+import type { EmptyComp, GameObj, OpacityComp } from "../../types";
+
+/**
+ * The {@link lifespan `lifespan()`} component.
+ *
+ * @group Components
+ */
+export interface LifespanCompOpt {
+    /**
+     * Fade out duration (default 0 which is no fade out).
+     */
+    fade?: number;
+}
 
 export function lifespan(time: number, opt: LifespanCompOpt = {}): EmptyComp {
     const k = getKaboomContext(this);
