+++ src/components/misc/stay.ts
@@ -1,4 +1,20 @@
-import type { StayComp } from "../../types";
+import type { Comp } from "../../types";
+
+/**
+ * The {@link stay `stay()`} component.
+ *
+ * @group Components
+ */
+export interface StayComp extends Comp {
+    /**
+     * If the obj should not be destroyed on scene switch.
+     */
+    stay: boolean;
+    /**
+     * Array of scenes that the obj will stay on.
+     */
+    scenesToStay: string[];
+}
 
 export function stay(scenesToStay?: string[]): StayComp {
     return {
