+++ src/components/transform/anchor.ts
@@ -1,5 +1,17 @@
 import type { Vec2 } from "../../math";
-import type { Anchor, AnchorComp } from "../../types";
+import type { Anchor, Comp } from "../../types";
+
+/**
+ * The {@link anchor `anchor()`} component.
+ *
+ * @group Components
+ */
+export interface AnchorComp extends Comp {
+    /**
+     * Anchor point for render.
+     */
+    anchor: Anchor | Vec2;
+}
 
 export function anchor(o: Anchor | Vec2): AnchorComp {
     if (!o) {
