+++ src/components/transform/fixed.ts
@@ -1,4 +1,16 @@
-import type { FixedComp } from "../../types";
+import type { Comp } from "../../types";
+
+/**
+ * The {@link fixed `fixed()`} component.
+ *
+ * @group Componens
+ */
+export interface FixedComp extends Comp {
+    /**
+     * If the obj is unaffected by camera
+     */
+    fixed: boolean;
+}
 
 export function fixed(): FixedComp {
     return {
