+++ src/components/transform/move.ts
@@ -1,5 +1,6 @@
 import { Vec2 } from "../../math";
-import type { EmptyComp, GameObj, PosComp } from "../../types";
+import type { EmptyComp, GameObj } from "../../types";
+import type { PosComp } from "./pos";
 
 export function move(dir: number | Vec2, speed: number): EmptyComp {
     const d = typeof dir === "number" ? Vec2.fromAngle(dir) : dir.unit();
