+++ src/components/transform/pos.ts
@@ -1,6 +1,71 @@
-import { getInternalContext, getKaboomContext } from "../../kaboom";
+import { getKaboomContext } from "../../kaboom";
 import { Vec2, vec2 } from "../../math";
-import type { FixedComp, GameObj, PosComp, Vec2Args } from "../../types";
+import type { Comp, GameObj, Vec2Args } from "../../types";
+import type { FixedComp } from "./fixed";
+
+/**
+ * The {@link pos `pos()`} component.
+ *
+ * @group Components
+ */
+export interface PosComp extends Comp {
+    /**
+     * Object's current world position.
+     */
+    pos: Vec2;
+    /**
+     * Move how many pixels per second. If object is 'solid', it won't move into other 'solid' objects.
+     */
+    move(xVel: number, yVel: number): void;
+    move(vel: Vec2): void;
+    /**
+     * Move how many pixels, without multiplying dt, but still checking for 'solid'.
+     */
+    moveBy(dx: number, dy: number): void;
+    moveBy(d: Vec2): void;
+    /**
+     * Move to a spot with a speed (pixels per second), teleports if speed is not given.
+     */
+    moveTo(dest: Vec2, speed?: number): void;
+    moveTo(x: number, y: number, speed?: number): void;
+    /**
+     * Get the position of the object on the screen.
+     */
+    screenPos(): Vec2;
+    /**
+     * Get the position of the object relative to the root.
+     */
+    worldPos(): Vec2;
+    /**
+     * Transform a local point (relative to this) to a screen point (relative to the camera)
+     */
+    toScreen(this: GameObj<PosComp | FixedComp>, p: Vec2);
+    /**
+     * Transform a local point (relative to this) to a world point (relative to the root)
+     * @since v3001.0
+     */
+    toWorld(this: GameObj<PosComp>, p: Vec2);
+    /**
+     * Transform a screen point (relative to the camera) to a local point (relative to this)
+     * @since v3001.0
+     */
+    fromScreen(this: GameObj<PosComp | FixedComp>, p: Vec2);
+    /**
+     * Transform a world point (relative to the root) to a local point (relative to this)
+     * @since v3001.0
+     */
+    fromWorld(this: GameObj<PosComp>, p: Vec2);
+    /**
+     * Transform a point relative to this to a point relative to other
+     * @since v3001.0
+     */
+    toOther(this: GameObj<PosComp>, other: GameObj<PosComp>, p: Vec2);
+    /**
+     * Transform a point relative to other to a point relative to this
+     * @since v3001.0
+     */
+    fromOther(this: GameObj<PosComp>, other: GameObj<PosComp>, p: Vec2);
+}
 
 function isFixed(obj: GameObj) {
     if (obj.fixed) return true;
@@ -9,7 +74,7 @@ function isFixed(obj: GameObj) {
 
 export function pos(...args: Vec2Args): PosComp {
     const k = getKaboomContext(this);
-    const internal = getInternalContext(k);
+    const { getViewportScale } = k._k;
 
     return {
         id: "pos",
@@ -117,7 +182,7 @@ export function pos(...args: Vec2Args): PosComp {
         drawInspect() {
             k.drawCircle({
                 color: k.rgb(255, 0, 0),
-                radius: 4 / internal.getViewportScale(),
+                radius: 4 / getViewportScale(),
             });
         },
     };
