+++ src/components/transform/rotate.ts
@@ -1,4 +1,26 @@
-import type { RotateComp } from "../../types";
+import type { Comp } from "../../types";
+
+/**
+ * The {@link rotate `rotate()`} component.
+ *
+ * @group Components
+ */
+export interface RotateComp extends Comp {
+    /**
+     * Angle in degrees.
+     */
+    angle: number;
+    /**
+     * Rotate in degrees.
+     */
+    rotateBy(angle: number): void;
+    /**
+     * Rotate to a degree (like directly assign to .angle)
+     *
+     * @since v3000.0
+     */
+    rotateTo(s: number): void;
+}
 
 export function rotate(r: number): RotateComp {
     return {
