+++ src/components/transform/scale.ts
@@ -1,11 +1,26 @@
-import { getInternalContext, getKaboomContext } from "../../kaboom";
-import { vec2 } from "../../math";
-import type { ScaleComp, Vec2Args } from "../../types";
+import { getKaboomContext } from "../../kaboom";
+import { type Vec2, vec2 } from "../../math";
+import type { Comp, Vec2Args } from "../../types";
+
+/**
+ * The {@link scale `scale()`} component.
+ *
+ * @group Components
+ */
+export interface ScaleComp extends Comp {
+    scale: Vec2;
+    scaleTo(s: number): void;
+    scaleTo(s: Vec2): void;
+    scaleTo(sx: number, sy: number): void;
+    scaleBy(s: number): void;
+    scaleBy(s: Vec2): void;
+    scaleBy(sx: number, sy: number): void;
+}
 
 // TODO: allow single number assignment
 export function scale(...args: Vec2Args): ScaleComp {
     const k = getKaboomContext(this);
-    const internal = getInternalContext(k);
+    const { toFixed } = k._k;
 
     if (args.length === 0) {
         return scale(1);
@@ -20,9 +35,7 @@ export function scale(...args: Vec2Args): ScaleComp {
             this.scale.scale(vec2(...args));
         },
         inspect() {
-            return `(${internal.toFixed(this.scale.x, 2)}, ${
-                internal.toFixed(this.scale.y, 2)
-            })`;
+            return `(${toFixed(this.scale.x, 2)}, ${toFixed(this.scale.y, 2)})`;
         },
     };
 }
