+++ src/components/transform/z.ts
@@ -1,4 +1,16 @@
-import type { ZComp } from "../../types";
+import type { Comp } from "../../types";
+
+/**
+ * The {@link z `z()`} component.
+ *
+ * @group Components
+ */
+export interface ZComp extends Comp {
+    /**
+     * Defines the z-index of this game obj
+     */
+    z: number;
+}
 
 export function z(z: number): ZComp {
     return {
